Move posthog/tasks/test/test_org_usage_report.py clickhouse version inline
